Add files via upload
This code merge dtm and dsm files from defra and calculate canopy height model for Bromley and Croydon area. Some places have gaps but we cant deal with it because it is an issue of initial data. 
Also I tested the calculation of height of bromley buildings  -  it looks realistic. 
Now we need to decide with what buildings we want to compare height and what metrics we gonna use